# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDF Transaction WebAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transaction-page": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction pages",
                "description": "Get transaction pages",
                "operationId": "get_transactionPages",
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create transaction page",
                "description": "Create transaction page",
                "operationId": "create_transactionPage",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "page"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "integer"
                                    },
                                    "start_offset": {
                                        "type": "text"
                                    },
                                    "end_offset": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/api/transaction-page/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction page",
                "description": "Get transaction page",
                "operationId": "get_transactionPage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction page id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update transaction page",
                "description": "Update transaction page",
                "operationId": "update_transactionPage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction page id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "page": {
                                        "type": "text"
                                    },
                                    "start_offset": {
                                        "type": "text"
                                    },
                                    "end_offset": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete transaction page",
                "description": "Delete transaction page",
                "operationId": "delete_transactionPage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction page id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}